Manually revert 91b2584845c95626e3595de67a5e45ba0667bf7e
I vaguely remember a website for which I introduced that but can't find it right now.
That was some kind of bulletin board on *.pp.ru or *.msk.ru domain, only one
subdomain of which was blocked, and it broke on full proxifying.

This fixes '*.echo.msk.ru' Registry entry.